Move drive collection and item handler out of OneDrive
