+++ src/internal/m365/collection/drive/collection.go
@@ -1,5 +1,5 @@
-// Package onedrive provides support for retrieving M365 OneDrive objects
-package onedrive
+// Package drive provides support for retrieving M365 Drive objects
+package drive
 
 import (
 	"context"
@@ -16,7 +16,7 @@ import (
 	"github.com/alcionai/corso/src/internal/common/ptr"
 	"github.com/alcionai/corso/src/internal/data"
 	"github.com/alcionai/corso/src/internal/m365/graph"
-	"github.com/alcionai/corso/src/internal/m365/service/onedrive/metadata"
+	"github.com/alcionai/corso/src/internal/m365/collection/drive/metadata"
 	"github.com/alcionai/corso/src/internal/m365/support"
 	"github.com/alcionai/corso/src/internal/observe"
 	"github.com/alcionai/corso/src/pkg/backup/details"
