+++ src/internal/m365/collection/drive/folder_cache.go
@@ -1,4 +1,4 @@
-package onedrive
+package drive
 
 import (
 	"github.com/microsoftgraph/msgraph-sdk-go/models"
