+++ src/internal/m365/collection/drive/item_collector.go
@@ -1,4 +1,4 @@
-package onedrive
+package drive
 
 import (
 	"context"
