+++ src/internal/m365/collection/drive/item_test.go
@@ -1,4 +1,4 @@
-package onedrive
+package drive
 
 import (
 	"bytes"
