+++ src/internal/m365/collection/drive/metadata/consts.go
