+++ src/internal/m365/collection/drive/metadata/metadata.go
