+++ src/internal/m365/collection/drive/metadata/permissions.go
