+++ src/internal/m365/collection/drive/metadata/testdata/permissions.go
