+++ src/internal/m365/collection/drive/permission.go
@@ -1,4 +1,4 @@
-package onedrive
+package drive
 
 import (
 	"context"
@@ -11,7 +11,7 @@ import (
 
 	"github.com/alcionai/corso/src/internal/common/ptr"
 	"github.com/alcionai/corso/src/internal/data"
-	"github.com/alcionai/corso/src/internal/m365/service/onedrive/metadata"
+	"github.com/alcionai/corso/src/internal/m365/collection/drive/metadata"
 	"github.com/alcionai/corso/src/internal/version"
 	"github.com/alcionai/corso/src/pkg/logger"
 	"github.com/alcionai/corso/src/pkg/path"
@@ -76,7 +76,7 @@ func getCollectionMetadata(
 		metaName = metadata.DirMetaFileSuffix
 	}
 
-	meta, err := fetchAndReadMetadata(ctx, dc, metaName)
+	meta, err := FetchAndReadMetadata(ctx, dc, metaName)
 	if err != nil {
 		return metadata.Metadata{}, clues.Wrap(err, "collection metadata")
 	}
