+++ src/internal/m365/collection/drive/restore_caches.go
@@ -1,4 +1,4 @@
-package onedrive
+package drive
 
 import (
 	"context"
@@ -11,7 +11,7 @@ import (
 	"github.com/alcionai/corso/src/internal/common/idname"
 	"github.com/alcionai/corso/src/internal/common/ptr"
 	"github.com/alcionai/corso/src/internal/m365/graph"
-	"github.com/alcionai/corso/src/internal/m365/service/onedrive/metadata"
+	"github.com/alcionai/corso/src/internal/m365/collection/drive/metadata"
 	"github.com/alcionai/corso/src/pkg/services/m365/api"
 )
