+++ src/internal/m365/collection/drive/url_cache.go
@@ -1,4 +1,4 @@
-package onedrive
+package drive
 
 import (
 	"context"
