+++ src/internal/m365/collection/drive/url_cache_test.go
@@ -1,4 +1,4 @@
-package onedrive
+package drive
 
 import (
 	"context"
