+++ src/internal/m365/service/onedrive/export.go
@@ -7,7 +7,8 @@ import (
 	"github.com/alcionai/clues"
 
 	"github.com/alcionai/corso/src/internal/data"
-	"github.com/alcionai/corso/src/internal/m365/service/onedrive/metadata"
+	"github.com/alcionai/corso/src/internal/m365/collection/drive"
+	"github.com/alcionai/corso/src/internal/m365/collection/drive/metadata"
 	"github.com/alcionai/corso/src/internal/version"
 	"github.com/alcionai/corso/src/pkg/backup/details"
 	"github.com/alcionai/corso/src/pkg/control"
@@ -120,7 +121,7 @@ func getItemName(
 		trimmedName := strings.TrimSuffix(id, metadata.DataFileSuffix)
 		metaName := trimmedName + metadata.MetaFileSuffix
 
-		meta, err := fetchAndReadMetadata(ctx, fin, metaName)
+		meta, err := drive.FetchAndReadMetadata(ctx, fin, metaName)
 		if err != nil {
 			return "", clues.Wrap(err, "getting metadata").WithClues(ctx)
 		}
